Pass "-lm" after object files, not before
This causes some undefined symbol errors on a machine I was trying to
build this on.  IIRC this is a pretty common way of doing things,
here's a reference:

http://www.linuxtopia.org/online_books/an_introduction_to_gcc/gccintro_18.html

This only fixes the problem for Linux-*-GCC, as it's the only
platform I have actually built this on.  The correct thing to do is
probably to use LDFLAGS like normal, but I don't want to change the
build system too much.

I've only tested this on Linux-x86_64-GCC, but the Makefile patches
applied to the other Linux-*-GCC targets so I figure it's probably OK.